+++ src/main.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import "leaflet/dist/leaflet.css";
 import "./i18n";
 import "./styles/tokens.css";
 import "./styles/global.css";